fix: Update release.yml
Adds a conditional to verify that actual changes are made; if no changes are found, the commit and push steps are skipped. This change is intended to prevent failures of the release workflow after the actual changes have already been pushed (the only remaining changes are package-lock.json and a gh hash of some sort).