feat: Adds script to check for changes (#467)
* feat: Adds script to check for changes

Adds the first version of find-changes.sh, which detects changed or added files in subfolders of samples/; this will allow us to optimize testing and building by operating only on things that have changed.

* feat: Update release.yml

Adds debug statement to release.yml, but importantly ALSO removes the build-all statement, in anticipation that it's not coming back.